+++ Chrominsky.Utils/Enums/DatabaseEntityStatus.cs
@@ -1,27 +1,30 @@
+using System.Text.Json.Serialization;
+
 namespace Chrominsky.Utils.Enums;
 
 /// <summary>
-/// Represents the status of a promo code.
+/// Represents the status of a db entity.
 /// </summary>
+[JsonConverter(typeof(JsonStringEnumConverter))]
 public enum DatabaseEntityStatus
 {
     /// <summary>
-    /// The promo code is active and can be used.
+    /// The db entity is active and can be used.
     /// </summary>
     Active,
 
     /// <summary>
-    /// The promo code is inactive and cannot be used.
+    /// The db entity is inactive and cannot be used.
     /// </summary>
     Inactive,
 
     /// <summary>
-    /// The promo code has been deleted and cannot be used.
+    /// The db entity has been deleted and cannot be used.
     /// </summary>
     Deleted,
 
     /// <summary>
-    /// The promo code is in draft mode and has not been published yet.
+    /// The db entity is in draft mode and has not been published yet.
     /// </summary>
     Draft,
 }