+++ Chrominsky.Utils/Models/Base/BaseSimpleEntity.cs
@@ -28,6 +28,6 @@ public class BaseSimpleEntity : IBaseDatabaseEntity
 
     /// <inheritdoc />
     [Required]
-    [Newtonsoft.Json.JsonConverter(typeof(JsonStringEnumConverter))]
+    [System.Text.Json.Serialization.JsonConverter(typeof(JsonStringEnumConverter))]
     public DatabaseEntityStatus Status { get; set; }
 }